winebow/requirements.txt
Django==3.2.13
django-allauth==0.50.0
mysqlclient==2.0.3
pyyaml==6.0
graphene-django==2.15.0
django-graphql-jwt==0.3.4
django-graphql-auth==0.3.16
validate-email==1.3
Pillow==9.1.0
django-summernote==0.8.20.0